requirements.txt
Django==1.8.4
PyYAML==3.11
diff-match-patch==20121119
django-import-export==0.2.8
django-suit==0.2.14
django-suit-locale==1.0.10
numpy==1.9.3
pandas==0.16.2
psycopg2==2.5.5
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
tablib==0.10.0
